[bootloader] Loading elf kernel works!
