+++ .vscode/settings.json
@@ -27,6 +27,7 @@
       "kmalloc.h": "c",
       "common_defs.h": "c",
       "term.h": "c",
-      "x86_utils.h": "c"
+      "x86_utils.h": "c",
+      "fat32.h": "c"
    }
 }
